Creating WordCountTokenizer Class (#802)
* feat: creating word count tokenizer class && falling back to word count tokenizer if tiktok fails

* feat: updating comment
